payments/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.7.1
pymongo==4.15.3
pydantic==2.7.0
strawberry-graphql==0.226.1